Reorganize programmatic usage files into dedicated folder
Move all programmatic usage tracking files into `lib/api/programmatic_usage/`
and drop the `programmatic_usage_` prefix from filenames for cleaner imports.

Files moved:
- daily_cap_tracking.ts → programmatic_usage/daily_cap.ts
- key_cap_tracking.ts → programmatic_usage/key_cap.ts
- programmatic_usage_common.ts → programmatic_usage/common.ts
- programmatic_usage_tracking.ts → programmatic_usage/tracking.ts
- programmatic_usage_tracking.test.ts → programmatic_usage/tracking.test.ts
