+++ front/lib/api/assistant/conversation.ts
@@ -28,9 +28,9 @@ import {
   publishMessageEventsOnMessagePostOrEdit,
 } from "@app/lib/api/assistant/streaming/events";
 import { maybeUpsertFileAttachment } from "@app/lib/api/files/attachments";
-import { getRemainingKeyCapMicroUsd } from "@app/lib/api/key_cap_tracking";
 import { getSupportedModelConfig } from "@app/lib/api/models";
-import { isProgrammaticUsage } from "@app/lib/api/programmatic_usage_tracking";
+import { getRemainingKeyCapMicroUsd } from "@app/lib/api/programmatic_usage/key_cap";
+import { isProgrammaticUsage } from "@app/lib/api/programmatic_usage/tracking";
 import type { Authenticator } from "@app/lib/auth";
 import { getFeatureFlags } from "@app/lib/auth";
 import { extractFromString } from "@app/lib/mentions/format";
