+++ front/pages/api/w/[wId]/keys/[id]/index.ts
@@ -3,7 +3,7 @@ import * as t from "io-ts";
 import type { NextApiRequest, NextApiResponse } from "next";
 
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
-import { invalidateKeyCapCache } from "@app/lib/api/key_cap_tracking";
+import { invalidateKeyCapCache } from "@app/lib/api/programmatic_usage/key_cap";
 import type { Authenticator } from "@app/lib/auth";
 import { KeyResource } from "@app/lib/resources/key_resource";
 import { apiError } from "@app/logger/withlogging";
